requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.13.1  # Database migrations
annotated-types==0.7.0
anthropic==0.69.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
# backports.asyncio.runner==1.2.0  # Not needed for Python 3.11+
bcrypt==4.3.0
beautifulsoup4==4.14.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.104.1
fastapi-limiter==0.1.6
filelock==3.19.1
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
limits==4.2
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
namecheap-python==0.2.3
numpy==2.0.2
openai==2.1.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.9  # Database adapter (PostgreSQL)
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.3
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==4.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==7.0.0
requests==2.32.4
requests-file==3.0.1
rsa==4.9.1
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==25.4.0
tldextract==5.3.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
xmltodict==1.0.2
yarl==1.20.1

# Production server
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Monitoring
prometheus-client==0.19.0

# TDS Auto-Sync Scheduler
apscheduler==3.11.1
python-socketio==5.11.1

# Testing factories and data generation
factory-boy==3.3.0  # Test data factories
Faker==30.8.2  # Fake data generation
pytest-mock==3.12.0  # Mock testing

# Schema comparison and migration
migra==3.0.1663481299  # Database schema comparison tool
